Update android files and add __tests__ directory from latest version of React Native v0.36
